backend/requirements.txt
Flask>=2.3.3
Flask-CORS>=4.0.0
pandas>=2.0.3
numpy>=1.24.3
python-dotenv>=1.0.0